layla: change base measure to mm and remove code narrow field
mm is better, because the outputs are meant to be physically printed.
for 2-d barcodes we can calculate the other line width fromn the first.

+++ layout.go
@@ -85,7 +85,7 @@ func (l *layouter) layout(o *Node, b Box, stack []*Node) error {
 			ca := a
 			cols := o.Cols
 			if len(cols) == 0 {
-				cols = []int{ca.W}
+				cols = []float64{ca.W}
 			}
 			ci := i % len(cols)
 			for _, cw := range cols[:ci] {
